requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohttp
azure-functions
azure-identity
azure-keyvault-secrets
beautifulsoup4
httpx
msgraph-sdk
mypy
pandas
yarl